Find homes on forest edge, set random pb
A small change made with the patch codes to identify patches on the forest edge that are habitat type = 0 and set random probability of ticks in those patches, but not in other hab_type = 0 patches. 

I have only run this with landscape 195 thus far, but will run with more landscape types to make sure the code functions as expected for each.